refactor: Improve Vulkan SDK installation and environment variable setup in CI workflow

+++ OloEngine/src/OloEngine/Renderer/Font.cpp
@@ -9,7 +9,8 @@
 #include "MSDFData.h"
 
 namespace OloEngine
-{	template<typename T, typename S, int N, msdf_atlas::GeneratorFunction<S, N> GenFunc>
+{
+	template<typename T, typename S, int N, msdf_atlas::GeneratorFunction<S, N> GenFunc>
 	static Ref<Texture2D> CreateAndCacheAtlas(const std::string_view /*fontName*/, f32 /*fontSize*/, const std::vector<msdf_atlas::GlyphGeometry>& glyphs,
 		const msdf_atlas::FontGeometry& /*fontGeometry*/, u32 width, u32 height)
 	{
@@ -100,14 +101,15 @@ namespace OloEngine
 		// if MSDF || MTSDF
 
 		u64 coloringSeed = 0;
-		bool expensiveColoring = false;		if (expensiveColoring)
+		bool expensiveColoring = false;
+		if (expensiveColoring)
 		{
 			msdf_atlas::Workload([&glyphs = m_Data->Glyphs, &coloringSeed](int i, int /*threadNo*/) -> bool
 			{
 				unsigned long long glyphSeed = coloringSeed ? ((LCG_MULTIPLIER * (coloringSeed ^ i)) + LCG_INCREMENT) : 0;
 				glyphs[i].edgeColoring(msdfgen::edgeColoringInkTrap, DEFAULT_ANGLE_THRESHOLD, glyphSeed);
 				return true;
-			}, m_Data->Glyphs.size()).finish(THREAD_COUNT);
+			}, static_cast<int>(m_Data->Glyphs.size())).finish(THREAD_COUNT);
 		}
 		else
 		{
